smart-food-ordering-platform/api-gateway/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
grpcio==1.60.0
grpcio-tools==1.60.0
pyjwt==2.8.0
python-dotenv==1.0.0
